requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
lightgbm
pyyaml
requests
python-dotenv
prophet
